requirements.txt
keras==3.4.1
matplotlib==3.8.2
numpy==2.1.0
pandas==2.2.2
python_dateutil==2.8.2
scikit_learn==1.5.1
streamlit==1.30.0
yfinance==0.2.41
